src/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["sarvensis <ivank_ri@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
pydantic-settings = "^2.1.0"
aiohttp = "^3.9.0"
elasticsearch = "^8.11.0"
eng-to-ipa = "0.0.2"
symspellpy = "6.7.7"
fasttext = { git = "https://github.com/cfculhane/fastText" }
pandas = "^2.1.3"
faiss-cpu = "1.7.4"
fastparquet = "^2023.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
